dev-requirements.txt
beautifulsoup4==4.6.0
Django==2.0.2
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-livereload-server==0.2.3
djangorestframework==3.7.7
icalendar==4.0.1
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
tornado==4.5.3
django-extensions==2.0.5
XlsxWriter==0.7.3
djangorestframework-simplejwt
django-cors-headers
